pins are being dropped live
